requirements.txt
fastapi[standard]
uvicorn
SQLAlchemy
python-dotenv
psycopg2-binary
python-jose
bcrypt==4.0.1
passlib[bcrypt]==1.7.4